requirements.txt
numpy == 2.2.3
pandas == 2.2.3
pillow == 11.3.0
opencv-python-headless==4.12.0.88
fastapi == 0.116.1
uvicorn == 0.35.0
Jinja2 == 3.1.6
python-multipart == 0.0.20
pyzbar == 0.1.9
easyocr == 1.7.2
groq == 0.30.0
langchain == 0.3.26
langchain_groq == 0.3.6
matplotlib == 3.10.3
paddleocr == 2.6.2
paddlepaddle == 2.9.1